fix: revamp command line argument parsing behavior
The clap configuration in the struct holding the application's options
was revamped to both better reflect documentation changes as well as fix
the faulty behavior of the the value parser for the model field.

The API key naming changes in the README are now reflected in the --help
command documentation.

The environment variables are now correctly configured to be read instead
of the command option.

The value parser function for the model field was refactored to use a
non-authorizing request, as it was discovered the API didn't required a
bearer token header type to check for the list of models. This also
allowed fixing the overflow issue when checking the api_key field upon
instantiation of a new cli struct parser, as now the value passed by clap
is the only required binding.