fix: add remaining extra time out behavior
A new error was introduced to handle cases of multiple responses with the
same empty content as a response. The error wasn't being thrown with the
correct message as it wasn't being processed beyond the return call of
the source site.

Now the error is also considered in the function designated for error
handling.